CXP-1418: custom app regenerate button acl tests
