+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/ConnectedApp/ConnectedAppSettings.tsx
@@ -13,7 +13,6 @@ type Props = {
 };
 
 export const ConnectedAppSettings: FC<Props> = ({connectedApp, monitoringSettings, handleSetMonitoringSettings}) => {
-
     return (
         <>
             <ConnectedAppMonitoringSettings
