+++ src/Akeneo/Connectivity/Connection/front/tests/src/connect/components/ConnectedApp/ConnectedAppSettings.test.tsx
@@ -67,11 +67,7 @@ test('Connected App Settings renders monitoring settings and authorizations', ()
     ).not.toBeInTheDocument();
 });
 
-test('Connected App Settings renders custom app credentials', () => {
-    const isGranted = jest.fn((acl: string) => {
-        return acl === 'akeneo_connectivity_connection_manage_test_apps';
-    });
-
+test('Connected Custom App Settings renders credentials', () => {
     const connectedApp = {
         id: '12345',
         name: 'Custom App A',
@@ -99,13 +95,11 @@ test('Connected App Settings renders custom app credentials', () => {
     });
 
     renderWithProviders(
-        <SecurityContext.Provider value={{isGranted}}>
-            <ConnectedAppSettings
-                connectedApp={connectedApp}
-                monitoringSettings={monitoringSettings}
-                handleSetMonitoringSettings={handleSetMonitoringSettings}
-            />
-        </SecurityContext.Provider>
+        <ConnectedAppSettings
+            connectedApp={connectedApp}
+            monitoringSettings={monitoringSettings}
+            handleSetMonitoringSettings={handleSetMonitoringSettings}
+        />
     );
 
     expect(
@@ -127,72 +121,4 @@ test('Connected App Settings renders custom app credentials', () => {
             exact: false,
         })
     ).toBeInTheDocument();
-});
-
-test('Connected App Settings does not renders custom app credentials when missing acl', () => {
-    const isGranted = jest.fn((acl: string) => {
-        return acl !== 'akeneo_connectivity_connection_manage_test_apps';
-    });
-
-    const connectedApp = {
-        id: '12345',
-        name: 'Custom App A',
-        scopes: ['scope 1'],
-        connection_code: 'some_connection_code',
-        logo: 'https://marketplace.akeneo.com/sites/default/files/styles/extension_logo_large/public/extension-logos/akeneo-to-shopware6-eimed_0.jpg?itok=InguS-1N',
-        author: 'Author A',
-        user_group_name: 'app_123456abcde',
-        connection_username: 'Connection Username',
-        categories: ['e-commerce', 'print'],
-        certified: false,
-        partner: null,
-        is_custom_app: true,
-        is_pending: false,
-        has_outdated_scopes: false,
-    };
-
-    const monitoringSettings = {
-        flowType: FlowType.DATA_DESTINATION,
-        auditable: true,
-    };
-
-    const handleSetMonitoringSettings = jest.fn(() => {
-        return;
-    });
-
-    renderWithProviders(
-        <SecurityContext.Provider value={{isGranted}}>
-            <ConnectedAppSettings
-                connectedApp={connectedApp}
-                monitoringSettings={monitoringSettings}
-                handleSetMonitoringSettings={handleSetMonitoringSettings}
-            />
-        </SecurityContext.Provider>
-    );
-
-    expect(
-        screen.queryByText('akeneo_connectivity.connection.connect.connected_apps.edit.settings.authorizations.title', {
-            exact: false,
-        })
-    ).toBeInTheDocument();
-
-    expect(
-        screen.queryByText('akeneo_connectivity.connection.connect.connected_apps.edit.settings.monitoring.title', {
-            exact: false,
-        })
-    ).toBeInTheDocument();
-
-    expect(screen.queryByText('ConnectedAppAuthentication')).toBeInTheDocument();
-
-    expect(
-        screen.queryByText('akeneo_connectivity.connection.connect.connected_apps.edit.settings.credentials.title', {
-            exact: false,
-        })
-    ).not.toBeInTheDocument();
-
-    expect(
-        screen.queryByText('akeneo_connectivity.connection.connect.connected_apps.edit.settings.credentials.title', {
-            exact: false,
-        })
-    ).not.toBeInTheDocument();
 });
